app/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
requests==2.25.1
PyYAML==5.3.1
cryptography==38.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
SQLAlchemy==1.4.15
pytest==7.1.2
pytest-cov==3.0.0
